test: split large test files per scenario (#496)
Two long-running test modules were flagged by the 2026-04-18 architecture
audit. Both are purely mechanical splits with no behavior changes — every
assertion was preserved.

apps/pipeline/tests/unit/test_rag.py (663 lines) →
  _rag_shared.py        shared make_chunk + parse_sse helpers
  test_rag_helpers.py   _format_timestamp, build_prompt, chunks_to_sources
  test_rag_endpoint.py  /api/ask SSE endpoint wiring
  test_rag_retrieval.py retrieve_chunks DB query builder
  test_rag_models.py    check_model_available
  test_rag_streaming.py Fireworks + Ollama streams + provider routing

apps/web/tests/unit/EpisodesList.test.tsx (551 lines) →
  EpisodesList.fixtures.ts     makeEpisode factory + mockEpisodes preset
  EpisodesList.display.test.tsx display / tag / tooltip / provider tests
  EpisodesList.filter.test.tsx  search / filter / clear / count tests

Pipeline: 33 tests pass (5 files, each < 265 lines).
Web: 285 tests pass (15 EpisodesList tests across 2 files, each < 140 lines).

Closes #496.
